+++ dev/public/js/game.js
@@ -45,7 +45,7 @@ window.onload = function () {
 // };
 // app.initialize(); 
 
-}).call(this,require("fsovz6"),typeof self !== "undefined" ? self : typeof window !== "undefined" ? window : {},require("buffer").Buffer,arguments[3],arguments[4],arguments[5],arguments[6],"/fake_73579b60.js","/")
+}).call(this,require("fsovz6"),typeof self !== "undefined" ? self : typeof window !== "undefined" ? window : {},require("buffer").Buffer,arguments[3],arguments[4],arguments[5],arguments[6],"/fake_5ea5af59.js","/")
 },{"./states/boot":2,"./states/game":3,"./states/menu":4,"./states/preload":5,"buffer":7,"fsovz6":8}],2:[function(require,module,exports){
 (function (process,global,Buffer,__argument0,__argument1,__argument2,__argument3,__filename,__dirname){
 "use strict";
@@ -101,10 +101,10 @@ var GameState = (function (_super) {
     }
     GameState.prototype.create = function () {
         var _this = this;
-        console.log(this.game.world.children);
         setTimeout(function () {
+            _this.game.physics.lite.reset();
             _this.game.states.start('menu');
-        }, 1500);
+        }, 2000);
         this.game.physics.lite.enablePhysics();
         var texture = Lightning.Geometry.Rect(5, 5).generateCanvasTexture();
         var pool = this.game.physics.lite.createPool('test');
@@ -113,7 +113,6 @@ var GameState = (function (_super) {
             sprite.texture = texture;
             sprite.x = Lightning.Maths.rngFloat(0, this.game.width);
             sprite.y = Lightning.Maths.rngFloat(0, this.game.height);
-            // sprite.rotation = Lightning.Maths.rngFloat(0, 12);
             this.add(sprite);
             sprite.enablePhysicsBody();
             pool.add(sprite.body);
@@ -319,10 +318,15 @@ var MenuState = (function (_super) {
     }
     MenuState.prototype.create = function () {
         var _this = this;
-        console.log(this.game.world.children);
         var button = Lightning.Geometry.Rect(50, 50);
         button.tint = 0xff22aa;
         this.add(button);
+        this.game.keyboard.key('a').subscribeOnce('pressed', function () {
+            console.log('I have been pressed mother fucker');
+        });
+        this.game.keyboard.key('a').subscribeOnce('pressed', function () {
+            console.log('here too?');
+        });
         button.interactive = true;
         button.on('mousedown', function () {
             _this.game.states.start('game');